+++ src/main/java/io/hhplus/concert/application/payment/PaymentDto.java
@@ -1,0 +1,17 @@
+package io.hhplus.concert.application.payment;
+
+import java.util.Date;
+
+import io.hhplus.concert.domain.payment.Payment;
+import lombok.Getter;
+
+@Getter
+public class PaymentDto {
+    long totalPrice;
+    Date paidAt;
+
+    public PaymentDto(Payment payment) {
+        this.totalPrice = payment.getPrice();
+        this.paidAt = payment.getPaidAt();
+    }
+}
