+++ src/main/java/io/hhplus/concert/application/waiting/WaitingDto.java
@@ -1,0 +1,19 @@
+package io.hhplus.concert.application.waiting;
+
+import io.hhplus.concert.domain.waiting.WaitingToken;
+import lombok.Getter;
+
+@Getter
+public class WaitingDto {
+    String token;
+    String status;
+    long waitingId;
+    long waitingAhead;
+
+    public WaitingDto(WaitingToken token, long waitingAhead) {
+        this.token = token.getToken();
+        this.status = token.getStatus().name();
+        this.waitingId = token.getId();
+        this.waitingAhead = waitingAhead;
+    }
+}
