--- src/main/java/io/hhplus/concert/facade/concert/ConcertScheduleDto.java
@@ -1,19 +1,0 @@
-package io.hhplus.concert.facade.concert;
-
-import java.util.Date;
-
-import io.hhplus.concert.domain.concert.ConcertSchedule;
-import lombok.Getter;
-
-@Getter
-public class ConcertScheduleDto {
-    Long id;
-    String place;
-    Date concertDate;
-
-    public ConcertScheduleDto(ConcertSchedule domain) {
-        this.id = domain.getId();
-        this.place = domain.getPlace();
-        this.concertDate = domain.getConcertDate();
-    }
-}
