--- src/main/java/io/hhplus/concert/facade/user/UserAssetFacade.java
@@ -1,22 +1,0 @@
-package io.hhplus.concert.facade.user;
-
-import org.springframework.stereotype.Component;
-
-import io.hhplus.concert.domain.user.UserAssetService;
-
-@Component
-public class UserAssetFacade {
-    UserAssetService userAssetService;
-
-    public UserAssetFacade(UserAssetService userAssetService) {
-        this.userAssetService = userAssetService;
-    }
-
-    public Long getBalance(long userId) {
-        return userAssetService.getUserAsset(userId).getBalance();
-    }
-
-    public Long chargeBalance(long userId, long amount) {
-        return userAssetService.chargeUserAsset(userId, amount).getBalance();
-    }
-}
