+++ src/main/java/io/hhplus/concert/interfaces/presentation/concert/dto/ConcertScheduleDto.java
@@ -1,0 +1,18 @@
+package io.hhplus.concert.interfaces.presentation.concert.dto;
+
+import java.util.Date;
+
+public class ConcertScheduleDto {
+    public static record Request(
+        String token,
+        long concertId
+    ) {
+    }
+
+    public static record Response(
+        long id,
+        Date date,
+        String address
+    ) {
+    }
+}
