+++ src/main/java/io/hhplus/concert/interfaces/presentation/payment/dto/PaymentDto.java
@@ -1,0 +1,15 @@
+package io.hhplus.concert.interfaces.presentation.payment.dto;
+
+public class PaymentDto {
+    public static record Request(
+        String token,
+        long reservationId
+    ) {
+    }
+
+    public static record Response(
+        long paidAmount,
+        long balance
+    ) {
+    }
+}
