+++ src/main/java/io/hhplus/concert/interfaces/presentation/user/dto/AssetChargeDto.java
@@ -1,0 +1,13 @@
+package io.hhplus.concert.interfaces.presentation.user.dto;
+
+public class AssetChargeDto {
+    public static record Request(
+        long amount
+    ) {
+    }
+
+    public static record Response(
+        long balance
+    ) {
+    }
+}
