+++ src/main/java/io/hhplus/concert/interfaces/presentation/user/dto/AssetGetDto.java
@@ -1,0 +1,8 @@
+package io.hhplus.concert.interfaces.presentation.user.dto;
+
+public class AssetGetDto {
+    public static record Response(
+        long balance
+    ) {
+    }
+}
