+++ src/main/java/io/hhplus/concert/interfaces/presentation/waiting/dto/WaitingTokenDto.java
@@ -1,0 +1,10 @@
+package io.hhplus.concert.interfaces.presentation.waiting.dto;
+
+public class WaitingTokenDto {
+    public static record Response(
+        String waitingToken,
+        long waitingId,
+        long waitingAhead
+    ) {
+    }
+}
