--- src/main/java/io/hhplus/concert/presentation/concert/dto/ConcertScheduleSeatDto.java
@@ -1,17 +1,0 @@
-package io.hhplus.concert.presentation.concert.dto;
-
-public class ConcertScheduleSeatDto {
-    public static record Request(
-        String token,
-        long concertScheduleId
-    ) {
-    }
-
-    public static record Response(
-        long id,
-        String location,
-        boolean reserved,
-        long price
-    ) {
-    }
-}
