--- src/main/java/io/hhplus/concert/presentation/payment/dto/AssetChargeDto.java
@@ -1,13 +1,0 @@
-package io.hhplus.concert.presentation.payment.dto;
-
-public class AssetChargeDto {
-    public static record Request(
-        long amount
-    ) {
-    }
-
-    public static record Response(
-        long balance
-    ) {
-    }
-}
