--- src/main/java/io/hhplus/concert/presentation/payment/dto/AssetGetDto.java
@@ -1,8 +1,0 @@
-package io.hhplus.concert.presentation.payment.dto;
-
-public class AssetGetDto {
-    public static record Response(
-        long balance
-    ) {
-    }
-}
