--- src/main/java/io/hhplus/concert/presentation/reservation/dto/ReservationDto.java
@@ -1,30 +1,0 @@
-package io.hhplus.concert.presentation.reservation.dto;
-
-import java.util.Date;
-import java.util.List;
-
-public class ReservationDto {
-    public static record Request(
-        List<Long> seatIds
-    ) {
-    }
-
-    public static record Response(
-        long reservationId,
-        long paymentId,
-        long totalPrice,
-        Date reservedAt,
-        Date expireAt,
-        List<ReservedSeatInfo> reservedSeats
-    ) {
-
-    }
-
-    public static record ReservedSeatInfo(
-        long id,
-        String location,
-        long price
-    ) {
-
-    }
-}
