--- src/main/java/io/hhplus/concert/presentation/waiting/WaitingScheduler.java
@@ -1,20 +1,0 @@
-package io.hhplus.concert.presentation.waiting;
-
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
-
-import io.hhplus.concert.facade.waiting.WaitingFacade;
-
-@Component
-public class WaitingScheduler {
-    WaitingFacade waitingFacade;
-
-    public WaitingScheduler(WaitingFacade waitingFacade) {
-        this.waitingFacade = waitingFacade;
-    }
-
-    @Scheduled(fixedDelay = 1000)
-    void scheduleWaitings() {
-        waitingFacade.scheduleWaiting();
-    }
-}
