--- src/main/java/io/hhplus/concert/presentation/waiting/dto/WaitingTokenDto.java
@@ -1,10 +1,0 @@
-package io.hhplus.concert.presentation.waiting.dto;
-
-public class WaitingTokenDto {
-    public static record Response(
-        String waitingToken,
-        long waitingId,
-        long waitingAhead
-    ) {
-    }
-}
