+++ src/test/java/io/hhplus/concert/reservation/ReservationFacadeIntegTest.java
@@ -10,9 +10,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import io.hhplus.concert.application.reservation.ReservationFacade;
 import io.hhplus.concert.domain.waiting.TokenService;
 import io.hhplus.concert.domain.waiting.WaitingToken;
-import io.hhplus.concert.facade.reservation.ReservationFacade;
 
 @SpringBootTest
 public class ReservationFacadeIntegTest {
